Recorder: Fix #13, handle errors when starting new session
Handle errors and exceptions better when starting a new session.
When choosing targets file and results folder, check for validity and
avoid crash when the input is not valid, or when the user closed the
dialog box for some reason. Return to dialog until the input is correct
or until the user chose to exit the dialog nicely.