Recorder: Add "Continue existing session" button
Add a button in the UI. The new button allows the user to choose a
previously created results folder, and continue the session from the
same place it was last stopped.
Recorder will look in the chosen folder for trials.csv file, and will
load all the necessary information/stats into the program.
The user will be allowed to choose a different configuration than the
one the session was created with.