Recorder: Add an option to specify rc code error types
Add a text field in the configuration window, and allow the user to
insert the error types. Read the user input and insert the value to the
error combo-box. When the user tags a target as an error, read the
chosen value from the combo-box.
If the user leaves the text field empty, the default rc code for error
will be "Error".