Recorder: Add an option to play target sound file
When starting a session, add an option in the configuration window to
choose a folder for sound files.
If the folder exists, the program will try to play a sound file for
every target. The sound file name need to appear in a new column in
"targets.csv" input file, under the name "sound file name".
When the user choose sounds folder, the trial's starting trigger
changes: the trigger will not be button press ("play" button), and not
the first tablet touch as when working without sound files.