Recorder: Fix #36 - Flip X axis recorded values
Currently, When recording tablet writing, the recorded points are
mirrored in respect to the X-axis only.
Fix this issue when recording points. Use the screen current resolution
and subtract the recorded X value.
(The tablet resolution in the X axis is determined by the screen
resolution, unlike the Y axis resolution which is fixed on 1020).