Python wrapper for wintab32 library.
Simple Wrapper for wintab32 library, designed to communicate with
tablets.
The wrapper contains the basic functions necessary for establishing a
connection with the tablet and poll it for packets.
wintab_params.py contains the necessary structures defined using ctypes.

The easiest way to poll the tablet for packets is to call
OpenTabletContext() and GetPackets().
The getTabletInfo() verifys there's a wintab supporting tablet connected.