Recorder: get tablet information using polling instead of events.
PyQt5 tabletEvents are limited to press/release/move events, and ignores
'hover' events (which are actually move events with pressure = 0).

Add a new QTimer() to run in the background, and on each timeout, poll
the tablet for packets, using wintab library wrapper. The sampling
happens every 5mS, but can be changed using the parameter
TABLET_POLL_TIME. The new tabletPoll function logic is very similar to
the tabletEvent function, with minor changes, such as checking
"manually" for press/release/move events.

When the WRecorder loads, use the wintab library to check if a supported
tablet is connect and open a tablet context.